requirements.txt
numpy>=1.19.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
Pillow>=8.0.0
matplotlib>=3.3.0
